[GREEN] get earthquake info happened between two dates from the USGS Earthquake Service

+++ domain/src/main/java/com/joseyustiz/earthquakeinfo/application/port/in/GetEarthquakeInfoUseCase.java
@@ -2,13 +2,14 @@ package com.joseyustiz.earthquakeinfo.application.port.in;
 
 import com.joseyustiz.earthquakeinfo.model.EarthquakeInfo;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
 public interface GetEarthquakeInfoUseCase {
     List<EarthquakeInfo> getInfoBetweenDates(LocalDate startDate, LocalDate endDate);
 
-    List<EarthquakeInfo> getInfoBetweenMagnitudes(double minMagnitude, double maxMagnitude);
+    List<EarthquakeInfo> getInfoBetweenMagnitudes(BigDecimal minMagnitude, BigDecimal maxMagnitude);
 
     List<EarthquakeInfo> getInfoBetweenTwoDateRanges(LocalDate startDateRange1, LocalDate endDateRange1, LocalDate startDateRange2, LocalDate endDateRange2);
